move base class to framework directory
